+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * dx: Clarity.
 * style: Exception messages.
 * feat: Uses transformer class names as service IDs for easier decoration.
+* refactor: Move mapping logic in `ObjectToObjectTransformer` to its own
+  service.
 
 ## 0.5.12